Link to selenium tests only when they are available.
It's not clear that _anything_ should show up when the self-diagnostics are
unavailable, but this is good enough for now.  Note that this uses
file_exists() on config.test.php.  This requires cooperation from
php's safe mode and the webserver config.  This should not be
a problem on any system that allows self-tests to be run.