Multiport: Keep PPA cookie per-port.
This completes the feature which allows multiple PPA instances
to run on a single web server on different ports -- excepting
the language cookie used by the Selenium self-tests.  If you
want to run different PPA instances on different ports of
a single server and want to run self-tests in different languages,
well, you're out of luck.