Remove jQuery 1.8 dependencies from js/database.js.
This also replaces window.setTimeout() with window.setInterval().

It turns out that using a html anchor as a container and
frobbing that with javascript is evil.  Eventually, your
last click will follow the anchor link, refresh the page,
and restart the process refresh -- even if what you clicked
on was supposed to stop refresh.   Replace this code with
an inline <div> and use <noscript> to make the non-js case
work.