Processes: Improve positioning of the animated loading graphic.
This moves the graphic from 3 em away from the control to 1 em, and adds
2 em of padding after the graphic.  Another option would be to move
the graphic to before the control, and pad with a transparent png
of the same size as the graphic to avoid bouncing when the loading
graphic is not displayed, or use another table cell, or whatever.
The loading graphic remains positioned on the right of the control
because that's where it's always been.